tests/wpt/tests/html/canvas/offscreen/canvas-host/2d.canvas.host.readonly.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.canvas.host.readonly</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.canvas.host.readonly</h1>
<p class="desc">Canvas objects are readonly</p>


<script>
var t = async_test("Canvas objects are readonly");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  var canvas2 = new OffscreenCanvas(100, 50);
  var d = ctx.canvas;
  _assertDifferent(canvas2, d, "canvas2", "d");
  ctx.canvas = canvas2;
  _assertSame(ctx.canvas, d, "ctx.canvas", "d");
  t.done();

});
</script>
