LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/canvas-host/2d.canvas.host.readonly-expected.txt

2d.canvas.host.readonly

canvas is readonly


PASS canvas is readonly
